Filtered out blogpost_comment_add in Action model
This method is confusing: you do not need to know Action to comment a blog post.
Pretending to know action AND blog post could lead to incoherent situation.
If there is no needing to know a detail to do something, don't ask for the detail.

left as #WAS: just to remember...