Tested 'ActionUpdateView' and 'ActionCreateView' views. Fixed some minor bugs.
Both views use the ActionForm form. However, there is an askbot view implementation
of the question update which uses EditQuestionForm and RevisionForm.
This means that a question update in askbot (performed whereas an Action is updated)
needs more computation with respect to the question creation.

Right now the Action update (seems to) works, but we should study that matter more deeply.