Problem with VoteManager usage on action votes.
It seems that 'self.question.votes' returns a RelatedManager on which the methods 'declareds' and 'anonymous' of (extended) VoteManager cannot be called.
An (ugly) hack colud be: 'Vote.objects.all() & self.question.votes.all()' ( A & B returns B if B is contained in A) returns directly a VoteQuerySet object.
This works, but but it probably could be done in a more elegant way (or maybe I'm simply misunderstanding things and the problem is different :) )